[ALPS04682295] Wlan: Data Stall and link quality
1. Sync parameters from customer.
2. Avoid build error for KASAN.

MTK-Commit-Id: 632d4cc0ba2e04e90c66103f0a2ea977ef5fe9bd

CR-Id: ALPS04682295
Feature:[Module]Wi-Fi Driver
(cherry picked from commit 6904de2891f74505c5b68fd7029b9522c152c562)

+++ common/wlan_lib.c
@@ -7368,8 +7368,7 @@ void wlanFinishCollectingLinkQuality(P_GLUE_INFO_T prGlueInfo)
 			  prLinkQualityInfo->u8LastTxFailCount) : 0;
 	if (u8TxTotalCntDif >= u8TxFailCntDif)
 		prLinkQualityInfo->u4CurTxPer = (u8TxTotalCntDif == 0) ? 0 :
-				(int)((float)(u8TxFailCntDif * 100) /
-					(float)(u8TxTotalCntDif));
+				(int)((u8TxFailCntDif * 100) / (u8TxTotalCntDif));
 	else
 		prLinkQualityInfo->u4CurTxPer = 0;
 	if (prLinkQualityInfo->u8IdleSlotCount < prLinkQualityInfo->u8LastIdleSlotCount) {
